ProyectosUNAB - v1.2.7/requirements.txt
django
django-filter
djangorestframework
gunicorn